clean production and phicpNP working
